requirements.txt
# PyTorch 2.7.0+ with CUDA support
# Note: For CUDA 12.6+, install PyTorch from the official site:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126
torch==2.7.0
torchvision
torchaudio

# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Image and video processing
opencv-python==4.10.0.84
numpy==2.1.1
pillow==11.0.0

# HTTP requests for video downloads
requests==2.32.3

# HuggingFace Transformers for SAM3
# Note: SAM3 requires transformers >= 4.45.0 (very recent)
# If not available, you may need to install from source or use the SAM3 repo directly
transformers>=4.45.0
accelerate>=0.30.0
huggingface_hub>=0.20.0

# Note: SAM3 is now loaded via HuggingFace Transformers
# Authentication required: huggingface-cli login

